Move local secret vault to API
