+++ src/API/Polyrific.Catapult.Api.SecretVault/Polyrific.Catapult.Api.SecretVault.csproj
@@ -2,10 +2,10 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.1</TargetFramework>
-	<RootNamespace>Polyrific.Catapult.Shared.SecretVault</RootNamespace>
+	<RootNamespace>Polyrific.Catapult.Api.SecretVault</RootNamespace>
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-    <PackageId>Polyrific.Catapult.Shared.SecretVault</PackageId>
-    <Description>Shared library for Polyrific Catapult Secret Vault</Description>
+    <PackageId>Polyrific.Catapult.Api.SecretVault</PackageId>
+    <Description>API Library for Polyrific Catapult Secret Vault</Description>
     <Version>1.0.1</Version>
     <Authors>frandi-polyrific</Authors>
     <Company>Polyrific, Inc</Company>
@@ -20,7 +20,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Polyrific.Catapult.Shared.Common\Polyrific.Catapult.Shared.Common.csproj" />
+    <ProjectReference Include="..\..\Shared\Polyrific.Catapult.Shared.Common\Polyrific.Catapult.Shared.Common.csproj" />
+    <ProjectReference Include="..\Polyrific.Catapult.Api.Core\Polyrific.Catapult.Api.Core.csproj" />
   </ItemGroup>
 
 </Project>
